Removes CometBFT (#3910). dagbft-integration's build is broken at HEAD and
this fixes it.

Also carries, from the review of this branch:

  - regenerated test/e2e2, whose stale empty import () failed the
    generated-files check
  - restored the simulator healing fixes the merge had silently reverted --
    the shared dispatcher (#4048) and HealInterval on both the executor and
    conductor options -- which had broken seven e2e tests plus test/validate
  - gosimports across 67 files
  - per-run netsim ports in the halt tests, ported from main

Merged with CI red. The remaining failures are pre-existing defects on
dagbft-integration, reproduced against the branch itself and tracked by
#4097: halt is never wired (nothing calls RegisterHaltController), anchor
recovery has no owner once Kourou retires the source-side push, and three
cmd/consensus-testnet tests fail. testreport.TestWriteOutputError fails only
in CI and passes locally.

+++ cmd/consensus-testnet/extended_test.go
@@ -286,7 +286,7 @@ func TestExtended_ThirtyMinuteIntegration(t *testing.T) {
 
 	// Check block heights
 	maxBlocks := uint64(0)
-	minBlocks := uint64(^uint64(0))
+	minBlocks := ^uint64(0)
 	for i := 0; i < numNodes; i++ {
 		blocks := executors[i].GetBlockCount()
 		t.Logf("Node %d blocks: %d", i, blocks)
